+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. 初始化 `sec watch` 命令
 2. `sec kline` 新增叠加指标
+3. 初始化 `sec announcements` 命令
 
 ### v0.3.6